Add tests to op_str

+++ src/operations/op_add.rs
@@ -30,7 +30,7 @@ pub fn op_add(register: &mut Register, instr: u16) {
     let imm_flag = (instr >> 5) & 0x1;
 
     if imm_flag == 1 {
-        let imm5 = sign_extend(instr & 0x1F, 5) as i16;
+        let imm5 = sign_extend(instr & 0x1F, 5);
         register.set(r0, (register.get(r1) as i16 + imm5) as u16);
     } else {
         let r2 = instr & 0x7;
