+++ src/operations/op_ldi.rs
@@ -23,7 +23,7 @@ use super::{sign_extend, update_flags};
 
 pub fn op_ldi(register: &mut Register, instr: u16, memory: &mut Memory) {
     let r0 = (instr >> 9) & 0x7;
-    let pc_offset = sign_extend(instr & 0x1FF, 9) as i16;
+    let pc_offset = sign_extend(instr & 0x1FF, 9);
     let addr = (register.pc as i16 + pc_offset) as u16;
     let indirect_addr = memory.read(addr);
     register.set(r0, memory.read(indirect_addr));
