+++ src/operations/op_sti.rs
@@ -10,7 +10,7 @@ use super::sign_extend;
 
 pub fn op_sti(register: &mut Register, instr: u16, memory: &mut Memory) {
     let r0 = (instr >> 9) & 0x7;
-    let pc_offset = sign_extend(instr & 0x1FF, 9) as i16;
+    let pc_offset = sign_extend(instr & 0x1FF, 9);
     let addr = memory.read((register.pc as i16 + pc_offset) as u16);
     memory.write(addr, register.get(r0));
 }
